requirements.txt
pandas
requests
sqlalchemy
PyYAML
psycopg2-binary
kaggle
python-dotenv
snowflake-connector-python
scikit-learn
